# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Northside Hospital Sales & Marketing Control Tower API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, completed, all]
            default: "all"
        - name: type
          in: query
          schema:
            type: string
            enum: [paid_search, social_media, content_marketing, direct_mail, community, email, all]
            default: "all"
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/top-performing:
    get:
      operationId: getTopCampaigns
      tags: [campaigns]
      summary: Get top performing campaigns
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 12m]
            default: "30d"
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignPerformance"
  /bookings/funnel:
    get:
      operationId: getBookingFunnel
      tags: [bookings]
      summary: Get patient booking funnel data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 12m]
            default: "30d"
        - name: serviceLineId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Funnel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingFunnel"
  /bookings/by-service-line:
    get:
      operationId: getBookingsByServiceLine
      tags: [bookings]
      summary: Get bookings grouped by service line
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 12m]
            default: "30d"
      responses:
        "200":
          description: Bookings by service line
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceLineBookings"
  /bookings/recent:
    get:
      operationId: getRecentBookings
      tags: [bookings]
      summary: Get recent patient bookings
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingSummary"
...
components:
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
        serviceLineId:
          type: integer
        serviceLineName:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
        spent:
          type: number
        impressions:
          type: integer
        clicks:
          type: integer
        conversions:
          type: integer
        ctr:
          type: number
        cpc:
          type: number
        cpa:
          type: number
        roi:
          type: number
      required:
        - id
        - name
        - type
        - status
        - budget
        - spent
        - impressions
        - clicks
        - conversions
    CampaignPerformance:
      type: object
      properties:
        campaignId:
          type: integer
        campaignName:
          type: string
        type:
          type: string
        roi:
          type: number
        conversions:
          type: integer
        spent:
          type: number
        cpa:
          type: number
      required:
        - campaignId
        - campaignName
        - type
        - roi
        - conversions
        - spent
        - cpa
    BookingFunnel:
      type: object
      properties:
        leads:
          type: integer
        appointmentsScheduled:
          type: integer
        appointmentsCompleted:
          type: integer
        followUps:
          type: integer
        leadsToScheduledRate:
          type: number
        scheduledToCompletedRate:
          type: number
        completedToFollowUpRate:
          type: number
        overallConversionRate:
          type: number
      required:
        - leads
        - appointmentsScheduled
        - appointmentsCompleted
        - followUps
        - leadsToScheduledRate
        - scheduledToCompletedRate
        - completedToFollowUpRate
        - overallConversionRate
    ServiceLineBookings:
      type: object
      properties:
        serviceLineId:
          type: integer
        serviceLineName:
          type: string
        totalBookings:
          type: integer
        newPatients:
          type: integer
        returningPatients:
          type: integer
        cancelledBookings:
          type: integer
        noShows:
          type: integer
        trend:
          type: number
      required:
        - serviceLineId
        - serviceLineName
        - totalBookings
        - newPatients
        - returningPatients
    BookingSummary:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: string
        serviceLineName:
          type: string
        leadSourceName:
          type: string
        status:
          type: string
        appointmentDate:
          type: string
        createdAt:
          type: string
        isNewPatient:
          type: boolean
      required:
        - id
        - patientId
        - serviceLineName
        - status
        - appointmentDate
        - createdAt
        - isNewPatient
...