# lib/api-zod/src/generated/types/bookingFunnel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Northside Hospital Sales & Marketing Control Tower API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingFunnel {
  leads: number;
  appointmentsScheduled: number;
  appointmentsCompleted: number;
  followUps: number;
  leadsToScheduledRate: number;
  scheduledToCompletedRate: number;
  completedToFollowUpRate: number;
  overallConversionRate: number;
}