# lib/api-zod/src/generated/types/campaignPerformance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Northside Hospital Sales & Marketing Control Tower API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignPerformance {
  campaignId: number;
  campaignName: string;
  type: string;
  roi: number;
  conversions: number;
  spent: number;
  cpa: number;
}