# lib/api-zod/src/generated/types/serviceLineBookings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Northside Hospital Sales & Marketing Control Tower API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceLineBookings {
  serviceLineId: number;
  serviceLineName: string;
  totalBookings: number;
  newPatients: number;
  returningPatients: number;
  cancelledBookings?: number;
  noShows?: number;
  trend?: number;
}